pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.0.3",
    "numpy>=2.1.2",
    "openai>=1.52.2",
    "pandas>=2.2.3",
    "plotly>=5.24.1",
    "pypdf2>=3.0.1",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "streamlit>=1.39.0",
    "textstat>=0.7.4",
    "python-magic==0.4.27",
    "python-magic-bin>=0.4.14",
    "pydantic==1.10.12"
]
